Split CreateAck and ManualClose components into seperate files
