+++ src/components/incident/CreateAckSignOffAction.tsx
@@ -1,0 +1,66 @@
+import React, { useState } from "react";
+
+// MUI
+import { MuiPickersUtilsProvider, KeyboardDatePicker } from "@material-ui/pickers";
+
+// Date io
+import DateFnsUtils from "@date-io/date-fns";
+
+// Api
+import { AcknowledgementBody } from "../../api";
+
+// Components
+import SignOffAction from "./SignOffAction";
+
+type CreateAckPropsType = {
+  onSubmitAck: (ack: AcknowledgementBody) => void;
+};
+
+const CreateAck: React.FC<CreateAckPropsType> = ({ onSubmitAck }: CreateAckPropsType) => {
+  const [selectedDate, setSelectedDate] = useState<Date | null>(null);
+
+  const handleSubmit = (msg: string) => {
+    // TODO: switch to use API when implemented in backend
+    onSubmitAck({
+      event: {
+        description: msg,
+        timestamp: new Date().toISOString(),
+      },
+      expiration: selectedDate && selectedDate.toISOString(),
+    });
+  };
+
+  const handleDateChange = (date: Date | null) => {
+    setSelectedDate(date);
+  };
+
+  return (
+    <SignOffAction
+      dialogTitle="Submit acknowledment"
+      dialogContentText="Write a message describing why this incident was acknowledged "
+      dialogSubmitText="Submit"
+      dialogCancelText="Cancel"
+      dialogButtonText="Create acknowledegment"
+      title="Submit acknowledment"
+      question="Are you sure you want to acknowledge this incident?"
+      onSubmit={handleSubmit}
+    >
+      <MuiPickersUtilsProvider utils={DateFnsUtils}>
+        <KeyboardDatePicker
+          disableToolbar
+          format="MM/dd/yyyy"
+          margin="normal"
+          id="expiry-date"
+          label="Expiry date"
+          value={selectedDate}
+          onChange={handleDateChange}
+          KeyboardButtonProps={{
+            "aria-label": "change date",
+          }}
+        />
+      </MuiPickersUtilsProvider>
+    </SignOffAction>
+  );
+};
+
+export default CreateAck;
